Move SDK retrofit creation into separate factory class
